Feature: User select player for roster spot
-Index.html: Added select button that pops open a modal where user
sees all players an can select which one they want and assign to
spot. Also added 'ids' to elements so knew which roster spot to
update.
-Index.js: Added event listener for select button and assign button
inside modal
-auth/events.js: updated route to correct function to run after user
signs in.
-player/api.js: updated ID so correct player is selected
-player/event.js: added 2 functions: 1 for when you open the select
modal it runs a get all players and displays that information inside
the modal so user can see what players are in the system. The second
function is when the modal is submitted with player and roster spot it
retrieves that specific players information.
-player/ui.js: 2 new UI functions: 1 to render all players inside of
modal and the second is to update the roster spot with the correct
persons information.