Updated Function Names to avoid Confusion
-Changed the name of two functions in players/events.js
-Changed onShowPlayersInitial to onShowPlayersOnSignIn to reflect when
it is being called, once you sign in this function runs to populate your
roster with your players.
-Changed onShowPlayersSelection to onShowPlayersForRosterSelectionModal
to describe when this is being called which is when you open
the selector modal to pick what players go to what roster spots this
shows you all available players to use.