Fixed Bug: Manger could assign not his players to roster
-Bug was a manager could assign players that were owned by other
managers to their roster.
-Api.js: update index function to be more versatile by updating
parameter input to ID
-Ui.js: removed console.log and extra parameter to allow function to
handle more calls
-Events.js: created new function that checked to make sure the
players owner was equal to the current user signed in. If it was it
passed along the response in the .then sequence if not it did
not pass the response of the API call so it breaks the sequence and the
error message is called saying you are trying to do a unlawful action